feat: migrate to signed types in BookmarkUiList and BookmarkBottomSheet

+++ presentation/src/main/java/com/london/presentation/feature/home/HomeScreen.kt
@@ -305,7 +305,7 @@ private fun Content(
                 BookmarkBottomSheet(
                     onSheetDismiss = homeScreenContract::onBookmarkSheetDismiss,
                     isSheetVisible = uiState.isBookmarkSheetVisible,
-                    bookmarkedMovieId = uiState.bookmarkedMovieId.toUInt()
+                    bookmarkedMovieId = uiState.bookmarkedMovieId
                 )
             }
         }
